chore(edit-content): avoid split methods in signal store
